Add inline embedded blobs for firmwares
This PR adds the functionality to compile the binary blob of each firmware that is used (not all existing ones) with the modules that use it. This increases the module size, but only a bit. As a benefit no firmware files or kernel support is needed to load the module. Also, this change does not remove the option to load non-embedded firmwares. The functionality is still there. It's only necessary to delete the firmware file before compiling, and provide the normal file inside /lib/modules afterwards.

The use cases for this:
- When the kernel has dynamic firmware loading disabled.
- When you want to use the kernel modules independently.
- When you don't want to distribute additional firmwares.

The patch works using three different scripts:
- firmware_fw.pl: Prepares the blob from .fw files by generating .h files.
- firmware_header.pl: Patches the .h files that load the firmwares if the blob exists.
- firmware_source.pl: Patches the .c files to use the new functions request/release_firmware().

Fully tested and reliable.